requirements.txt
streamlit
easyocr
pandas
scikit-learn
opencv-python-headless
numpy
pillow
